Absolutely position renderer to reduce container size
The container adds approx 7px to the height of child elements which makes the page scroll. I can't find where it is coming from (padding 0, margin 0, border: 0, etc...). Taking the renderer out of document flow resolves this as the container has a height of 0 and so don't make the page scroll.